프로젝트2(코드정리)/requirements.txt
Flask
Flask-WTF
requests-oauthlib
python-dotenv